futility/futility-py/pyproject.toml
[project]
name = "futility-py"
version = "0.1.0"
description = ""
authors = []
requires-python = ">=3.13"
dependencies = [
    "gitpython (>=3.1.44,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true

[tool.poetry]
package-mode=false

[tool.pyright]
venvPath = "."
venv = ".venv"